+++ src/main/java/com/redhat/devtools/lsp4ij/commands/editor/TriggerSuggestAction.java
@@ -10,52 +10,17 @@
  ******************************************************************************/
 package com.redhat.devtools.lsp4ij.commands.editor;
 
-import com.intellij.openapi.actionSystem.*;
-import com.intellij.openapi.actionSystem.ex.ActionUtil;
-import com.intellij.openapi.actionSystem.impl.SimpleDataContext;
-import com.intellij.openapi.fileEditor.FileEditor;
-import com.intellij.openapi.fileEditor.FileEditorManager;
-import com.intellij.openapi.fileEditor.TextEditor;
-import com.intellij.openapi.project.Project;
-import com.redhat.devtools.lsp4ij.commands.LSPCommand;
-import com.redhat.devtools.lsp4ij.commands.LSPCommandAction;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import com.intellij.openapi.actionSystem.ActionPlaces;
 
 /**
  * Emulates Visual Studio Code's "editor.action.triggerSuggest" command, to trigger code completion after selecting a completion item.
  */
-public class TriggerSuggestAction extends LSPCommandAction {
+public class TriggerSuggestAction extends TriggerCommandAction {
+
 
     private static final String CODE_COMPLETION_ACTION = "CodeCompletion";
 
-    @Override
-    protected void commandPerformed(@NotNull LSPCommand command, @NotNull AnActionEvent e) {
-        Project project = e.getProject();
-        if (project == null) {
-            return;
-        }
-        @Nullable TextEditor activeEditor = findActiveEditor(project);
-        if (activeEditor == null) {
-            return;
-        }
-        AnAction codeCompletionAction = ActionManager.getInstance().getAction(CODE_COMPLETION_ACTION);
-        if (codeCompletionAction != null ) {
-            DataContext dataContext = createDataContext(activeEditor, project);
-            ActionUtil.invokeAction(codeCompletionAction, dataContext, ActionPlaces.UNKNOWN, null, null);
-        }
-    }
-
-    private static DataContext createDataContext(TextEditor activeEditor, Project project) {
-        SimpleDataContext.Builder contextBuilder = SimpleDataContext.builder();
-        contextBuilder.add(CommonDataKeys.PROJECT, project)
-                      .add(CommonDataKeys.EDITOR, activeEditor.getEditor());
-        return contextBuilder.build();
-    }
-
-    public static @Nullable TextEditor findActiveEditor(Project project) {
-        FileEditorManager fileEditorManager = FileEditorManager.getInstance(project);
-        FileEditor editor = fileEditorManager.getSelectedEditor();
-        return editor instanceof TextEditor? (TextEditor) editor: null;
+    public TriggerSuggestAction() {
+        super(CODE_COMPLETION_ACTION, ActionPlaces.UNKNOWN);
     }
 }
