+++ src/main/resources/templates/lsp/luau-lsp/clientSettings.json
@@ -1,0 +1,1 @@
+{}
