+++ src/main/resources/templates/lsp/luau-lsp/settings.json
@@ -1,0 +1,1 @@
+{}
