+++ src/main/resources/templates/lsp/luau-lsp/settings.schema.json
@@ -1,0 +1,1 @@
+{}
